refactor: Voucher 타입 클래스들을 하나의 패키지로 이동
