+++ src/main/java/com/devcourse/voucherapp/entity/voucher/FixDiscountVoucher.java
@@ -1,4 +1,4 @@
-package com.devcourse.voucherapp.entity;
+package com.devcourse.voucherapp.entity.voucher;
 
 import java.util.UUID;
