+++ src/main/java/com/devcourse/voucherapp/service/VoucherService.java
@@ -1,6 +1,6 @@
 package com.devcourse.voucherapp.service;
 
-import com.devcourse.voucherapp.entity.Voucher;
+import com.devcourse.voucherapp.entity.voucher.Voucher;
 import com.devcourse.voucherapp.entity.VoucherType;
 import org.springframework.stereotype.Service;
